+++ src/main/kotlin/org/kalasim/ComponentGenerator.kt
@@ -6,7 +6,7 @@ import org.koin.core.Koin
 import org.koin.core.context.GlobalContext
 
 /**
- * A component generator can be used to generate components
+ * A component generator can be used to generate components.
  *
  * See https://www.salabim.org/manual/ComponentGenerator.html
  *
@@ -14,34 +14,41 @@ import org.koin.core.context.GlobalContext
  *  * according to a given inter arrival time (iat) value or distribution
  *  * random spread over a given time interval
  *
- *  @param from time where the generator starts time
- *  @param till time up to which components should be generated. If omitted, no end
+ *  @param iat Inter arrival time arrival time distribution between arrivals/generations.
+ *  @param startAt time where the generator starts time. If omitted, `now` is used.
+ *  @param forceStart If `false` (default), the first component will be generated at `time = startAt + iat()`. If `true`, the first component will be generated at `startAt`.
+ *  @param until time up to which components should be generated. If omitted, no end.
+ *  @param total (maximum) number of components to be generated.
+ *  @param name name of the component.  if the name ends with a period (.), auto serializing will be applied  if the name end with a comma, auto serializing starting at 1 will be applied  if omitted, the name will be derived from the class it is defined in (lowercased)
+ * @param priority If a component has the same time on the event list, this component is sorted according to the priority. An event with a higher priority will be scheduled first.
+ *  @param storeRefs If `true`, i will store a reference of all generated components which can be queried with `arrivals`.
+ * @param koin The dependency resolution context to be used to resolve the `org.kalasim.Environment`
  */
 class ComponentGenerator<T>(
     val iat: RealDistribution,
-    val from: Double? = 0.0,
-    var till: Double = Double.MAX_VALUE,
+    val startAt: Double? = null,
+    val forceStart: Boolean = false,
+    var until: Double = Double.MAX_VALUE,
     val total: Int = Int.MAX_VALUE,
     name: String? = null,
     priority: Priority = NORMAL,
-    val storeRefs : Boolean = false,
-    @Suppress("UNUSED_PARAMETER") urgent: Boolean = false,
+    val storeRefs: Boolean = false,
     koin: Koin = GlobalContext.get(),
     val builder: Environment.(counter: Int) -> T
-) : Component(name, priority = priority, process = ComponentGenerator<T>::doIat, koin = koin) {
+) : Component(name, priority = priority, at = startAt, process = ComponentGenerator<T>::doIat, koin = koin) {
 
-    fun  interface Consumer<K> {
+    fun interface Consumer<K> {
         fun consume(generated: K)
     }
 
-    private val consumers  = mutableListOf<Consumer<T>>()
+    private val consumers = mutableListOf<Consumer<T>>()
     fun addConsumer(consumer: Consumer<T>) = consumers.add(consumer)
     fun removeConsumer(consumer: Consumer<T>) = consumers.remove(consumer)
 
     init {
         // TODO build intervals
 
-        if(storeRefs){
+        if(storeRefs) {
             consumers.add(ArrivalTracker())
         }
     }
@@ -49,21 +56,23 @@ class ComponentGenerator<T>(
     fun doIat(): Sequence<Component> = sequence {
         var numGenerated = 0
 
-        while (true) {
-            val t = env.now + iat.sample()
+        val iatSeq = sequence { if(forceStart) yield(0.0); while(true) yield(iat()) }.iterator()
 
-            if (t > till) {
-                yield(activate(at = till, process = ComponentGenerator<T>::doFinalize))
+        while(true) {
+            val interArrivalTime = iatSeq.next()
+
+            if((env.now + interArrivalTime) > until) {
+                yield(activate(at = until, process = ComponentGenerator<T>::doFinalize))
             }
 
-            hold(till = t)
+            hold(interArrivalTime)
 
             val created = builder(env, numGenerated)
             numGenerated++
 
-            consumers.forEach{ it.consume(created)}
+            consumers.forEach { it.consume(created) }
 
-            if (numGenerated >= total) break
+            if(numGenerated >= total) break
         }
     }
 
@@ -76,7 +85,7 @@ class ComponentGenerator<T>(
 
     // https://www.baeldung.com/kotlin/observer-pattern
 
-    class  ArrivalTracker<T> : ArrayList<T>(), Consumer<T>{
+    class ArrivalTracker<T> : ArrayList<T>(), Consumer<T> {
         override fun consume(generated: T) {
             add(generated)
         }
@@ -88,4 +97,4 @@ class ComponentGenerator<T>(
 }
 
 
-class ComponentGeneratorInfo<T >(cg: ComponentGenerator<T>) : Component.ComponentInfo(cg)
+class ComponentGeneratorInfo<T>(cg: ComponentGenerator<T>) : Component.ComponentInfo(cg)